# contracts/strategy-signal-service.openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "strategy-signal-service",
    "description" : "Strategy registry + live signal engine (Stage C): D18 lifecycle, signals with score breakdowns",
    "version" : "v1"
  },
  "paths" : {
...
    "/api/v1/strategies/{id}/universe" : {
      "get" : {
        "tags" : [ "universe-controller" ],
        "operationId" : "universe",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "version",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UniverseInfo"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/insights/compare" : {
      "get" : {
        "tags" : [ "insight-controller" ],
        "operationId" : "compare",
        "parameters" : [ {
          "name" : "signalIds",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompareResult"
                }
              }
            }
          },
          "default" : {
            "description" : "Error envelope (COMMON 8.3)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Constituent" : {
        "type" : "object",
        "properties" : {
          "exchange" : {
            "type" : "string"
          },
          "tradingsymbol" : {
            "type" : "string"
          }
        },
        "required" : [ "exchange", "tradingsymbol" ]
      },
      "UniverseInfo" : {
        "type" : "object",
        "properties" : {
          "mode" : {
            "type" : "string"
          },
          "asOf" : {
            "type" : [ "string", "null" ]
          },
          "constituentCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "checksum" : {
            "type" : "string"
          },
          "survivorshipCaveat" : {
            "type" : [ "string", "null" ]
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Constituent"
            }
          }
        },
        "required" : [ "asOf", "checksum", "constituentCount", "items", "mode", "survivorshipCaveat" ]
      },
...
      "CompareColumn" : {
        "type" : "object",
        "properties" : {
          "signalId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "tradingsymbol" : {
            "type" : "string"
          },
          "side" : {
            "type" : "string"
          },
          "family" : {
            "type" : "string"
          },
          "book" : {
            "type" : "string"
          },
          "priority" : {
            "type" : [ "string", "null" ]
          },
          "band" : {
            "type" : [ "string", "null" ]
          },
          "components" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ComponentPoint"
            }
          },
          "optionLegCost" : {
            "type" : [ "string", "null" ]
          },
          "marginEstimate" : {
            "type" : "string"
          },
          "riskReward" : {
            "type" : [ "string", "null" ]
          },
          "entryPrice" : {
            "type" : [ "string", "null" ]
          },
          "stopLoss" : {
            "type" : [ "string", "null" ]
          },
          "target" : {
            "type" : [ "string", "null" ]
          },
          "dataTrust" : {
            "type" : "string"
          },
          "trustReasons" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "scored" : {
            "type" : "string"
          }
        },
        "required" : [ "band", "book", "components", "dataTrust", "entryPrice", "family", "marginEstimate", "optionLegCost", "priority", "riskReward", "scored", "side", "signalId", "stopLoss", "target", "tradingsymbol", "trustReasons" ]
      },
      "CompareResult" : {
        "type" : "object",
        "properties" : {
          "session" : {
            "type" : "string",
            "format" : "date"
          },
          "booksDiffer" : {
            "type" : "boolean"
          },
          "differsMost" : {
            "type" : [ "string", "null" ]
          },
          "columns" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CompareColumn"
            }
          },
          "notes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "booksDiffer", "columns", "differsMost", "notes", "session" ]
      },
      "ComponentPoint" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "points" : {
            "type" : [ "string", "null" ]
          },
          "c" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "c", "key", "points" ]
      },
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object"
          }
        }
      }
    }
  }
}